Incorrect handling of TAB key in SuggestionPopup
This is part of the fix for JIRA PIVOT-762.

Straightens out the handling of Tab/Shift-Tab keys in SuggestionPopup
(in the TerraSuggestionPopupSkin) by correctly checking for these keys
in the ListView rather than the SuggestionPopup itself, and then
setting a variable everywhere: "returnFocusToTextInput" which
determines whether the key or mouse click is trying to focus
elsewhere once the popup is closed, or whether (in the case of
ENTER or ESCAPE) the focus should go right back to the TextInput.

Note: there are still issues of whether ESCAPE should restore
the "previous" text or not and there are other components that
"pop up" that should have similar fixes for Tab/Shift-Tab handling.