Merging revision 1366521 from trunk to branches/2.0.x
PIVOT-860: CPU usage at 100% when displaying an Expander.

The TerraExpanderSkin.paint method was calling "setTitleBarColor" because of the
color change before and after the transition due to the enabled state changing.
This caused an infinite repaint loop that used 100% CPU and was exacerbating
a memory leak which would soon eat up all memory.

So, took out the set color in the paint method and implemented the
ComponentStateListener method "enabledChanged" which sets the title bar color
and avoids the infinite repaint loop.