Merging revision 1366562 from 2.0.x branch to trunk.
Pivot-783: ScaleDecorator doesn't repaint component when scale changes.

Added Javadoc to the ScaleDecorator class to document how to make the repaint
happen (by calling component.repaint() or removing and re-adding the decorator).

This is much easier than redoing the whole decorator infrastructure for this
one particular case.  See the discussion in JIRA.