PIVOT-873: Add a style to TableView to allow editing on first mouse click.
Implemented the "editOnMouseDown" style to TerraTableViewSkin to allow
the first mouse click to initiate row editing instead of waiting for
a double click.

Note: editing is initiated, but the mouse click is not propagated to
a child field (for instance, checkbox or radiobutton) in order to
actually select/toggle the control.

This is a merge of revision 1388236 from branches/2.0.x to trunk.