PIVOT-880: NPE when setting the root directory to a Windows UNC path in a FileBrowserSheet.
Caused by the fact that the \\ doesn't match any items in the drive
list, and one of the listeners assumes the item will always be there.
Simple test for a non-null item fixes the issue and allows the UNC
path to be used, and to display items.

This is a merge of revision 1407585 from trunk to branches/2.0.x.