PIVOT-887: Make changes to Spinner and BXMLSerializer such that using the DefaultProperty of Spinner (which is "spinnerData") in a BXML file will work correctly.
Normally if the original value is a Sequence (as it always will be for the spinner data) the child is added to the sequence.  But using an ImmutableList enables BXMLSerializer to catch the UnsupportedOperationException and do a beanAdapter.put instead (which results in a call to "setSpinnerData"), which is the correct thing to do.

Merge revision 1428650 from trunk to branches/2.0.x.