PIVOT-892: Floating-point validators were not recognizing scientific notation as valid input.
From a lot of online searching, and reading through the Java code, it
appears that scientific notation is only enabled in NumberFormat by
applying a pattern (i.e., there is no factory method that returns
a suitable NumberFormat).  And even then, exponents are only recognized
if the capital E is used (not lower-case 'e').  So, adjust for that in
both the default and locale-specific cases.

Updated a bit of documentation in the FormattedValidator class to make
things more clear.

This is a merge of revision 1444260 from trunk to branches/2.0.x.