Part of PIVOT-891: Implement double-click to select a word and triple-click to select the whole line.
Fix the Shift-LEFT and Shift-RIGHT to work "properly" in that case of first
moving in one direction and then reversing direction -- this should reduce the
selection that was just done instead of always expanding it.

The same (or similar) changes for UP/DOWN are more difficult and are eluding me
at the moment.

Also implement Ctrl-HOME moves to the beginning of the document and Ctrl-END to
the end of the document (and with Shift selects from cursor to the beginning/end).