PIVOT-927:  Seemingly random occurrence of a NullPointerException during mouse operations causes the UI thread to crash, making the application unresponsive.
From the stack trace the problem is a null location for the mouse which
is not checked before referencing the location variable.

So, protect all references to the location with a null check first.  Also
make some effort to use the display mouse location if necessary (that is,
if the "mapPointFromAncestor" returns a null location).