PIVOT-951: Only first click is received by container listeners
Only the first (ever) click into a container is received by listeners
because the "mouseClickConsumed" variable never gets reset after the
first click has been detected.  This change simply resets the variable
on a mouse up event, which enables all subsequent clicks to be
received as well.

It is noted that perhaps this variable is not really necessary and
could be removed. Will investigate further.

This is a merge of revision 1602230 from branches/2.0.x to trunk.