PIVOT-953: Update startup positioning code to make sure that the final position for a desktop application is within the current virtual screen area.  This is a better fix for earlier code that attempted to make the x and y values never go negative for two reasons:  the previous code only worked on the --x and --y values specified on the command line, but not for saved positions, and for multiple monitor configurations where the secondary monitor(s) are left / above the primary monitor, the actual virtual screen coordinates could legitimately be negative and the window is still visible.
So, compute the bounds of the whole virtual screen space, by getting the union
of all the screen device bounds, then make sure the x and y coordinates (just
before using them to set the main window position) are within the virtual screen
boundaries.

This is a merge of revision 1610563 from trunk to branches/2.0.x.