Code cleanup:  Reorganize the "build.xml" so that the tests are compiled after everything else, in case there are new methods that the tests depend on that won't compile until the core/wtk .jar files are built.
Add a copy of .bxml, .json and .png files in both the tests and tutorials
directories, so you can just change to the xxxx/ant-bin directory and run
tests with "java org.apache.pivot....." (that is the class name of the
test or tutorial).  All the resources will be in place to do that now.