PIVOT-964: Fix another place where the "nashorn.global" object has to be searched in order to find script variables defined when using Java 8.
The trick is that now script variables are a part of the Nashorn global
space, which is a sub-object of the regular namespace.  So, the additional
steps which are necessary in the case that a variable is not found would
be to then get the "nashorn.global" object, and if it exists transform
it into a "Bindings" (if possible) and then search for the value there.

Note: this fixes the particular case reported by user, but I'm pretty
sure there are other places that needs this same treatment.