PIVOT-986:  Provide a property for the FileBrowser and VFSBrowser classes to allow viewing of hidden files.
There is already a STYLE in the skins of these classes that is nicely
analogous, so make a new style in the skins called "showHiddenFiles",
provide getter and setter methods and then "do the math" inside the
skins to take out the HIDDEN_FILE_FILTER if this flag is set, while
still using it when the flag is false.

Provide a sort-of pass-through style in the SheetSkins to set the
style in the component skin (for convenience).  Again this is based
on the way the "hideDisabledFiles" style works (exactly).

That's it, really.  Except add a simple test program in VFSBrowserTest
(for the VFSBrowser, obviously) to prove that the code works.  I need
to find a similar suitable place to test the regular FileBrowser case,
though, which will be a separate submission.

The FileBrowser parts of this change will be crossed to the 2.0.x
branch in due course.