PIVOT-987:  Address the different handling of script return values from attributes versus elements.
The script return value was being ignored in AttributeInvocationHandler
and the default values were always being used.  But in the similar
code in ElementInvocationHandler the script result was being used
and the defaults only used if the script method returns nothing.

Bring the code for attributes in line with that for elements.