PIVOT-987:  Address the different handling of script return values from attributes versus elements.
The script return value was being ignored in AttributeInvocationHandler
and the default values were always being used.  But in the similar
code in ElementInvocationHandler the script result was being used
and the defaults only used if the script method returns nothing.

Bring the code for attributes in line with that for elements.

This is a merge of revision r1740570 from trunk to branches/2.0.x.

Note: the merge includes the minor change also from trunk that
changes the return value in the Boolean case from "false" to
Boolean.FALSE.  I believe this was found by a "lint" check of
the code.