PIVOT-891:  Add double/triple click support in TextPane the same way it is supported (now) in TextArea.
Double click selects the word immediately under/around the mouse pointer,
and triple click selects the line (paragraph) that is under the mouse.

Essentially we just duplicated the code from TextArea into TextPane with
appropriate (small) modifications for the difference in APIs.

Had to port over the "getRowOffset" and "getRowLength" methods also (which
were not part of the original change in "trunk" because they were added by
earlier changes).

Otherwise, this is a merge of revision 1750549 from "trunk" to "branches/2.0.x".