PIVOT-992:  Change the "getSelectedText" method of TextPane to just use the equivalent call to "getText(start, finish)" instead of doing a "getSelectedRange()" and serializing that.  The results will be different in that if the selection doesn't actually encompass an end of line, then the result will not contain a spurious newline.
The difference affects find/replace operations (for instance) that would be calling
"getSelectedText" and doing a find operation on a piece of text that doesn't actually
appear in the document (like "abc\n").

I suspect that this new method is actually a bit faster than the old way, although
that probably makes little to no difference in actual use.