PIVOT-993:  Update Version.java to work with Oracle JDK 8u131.
Problem related to trying to parse the "131" value into a byte
which results in overflow.

Take patch proposed by Todd Volkert and widen all the fields
in Version to "short", and the "getNumber()" result to a long
to account for the wider fields.

------------------------------------------------------------------------
PIVOT-993, update formatting string

------------------------------------------------------------------------
fix unnecessary widening of formatting

------------------------------------------------------------------------
This is a merge of revisions 1792480, 1792481 and 1792500 from
"branches/2.0.x" to "trunk".