PIVOT-850:  First phase of Input Method Editor support for the text controls: 1) Add the infrastructure to pass the message in/out of the text fields from    the DisplayHost (which is the AWT component where they get sent first). 2) Changes to TerraTextInputSkin to do all the drawing / hit testing, etc.    using TextLayout instead of GlyphVector. 3) Implement the required interfaces in TerraTextInputSkin, which are    subclassed from new classes. 4) Define the interfaces to deal with AttributedCharacterIterator and    InputMethodRequests (and also InputMethodListener).
There are some unresolved issues left having to do with cursor movement
when composed text is present.  These will be resolved in a later
submission.

Note also, the basic infrastructure is there, but only TextInput is
presently dealing with it.  TextArea and TextPane, which are more
difficult will come later.