PIVOT-850: Update the text width calculations to use an average of the TextLayout "bounds" and the "pixel bounds" which can differ by 4 pixels (often with the default font), and maybe more.  Using the average seems visually to be "all right", in that the caret doesn't look too weird.
Fix problems with the drawing of the prompt with centered or right-justified
controls.  We need to calculate the "alignmentDeltaX" with the prompt text
instead of the real text (which there isn't any of when drawing the prompt).

Add centered and right-justified controls to the "text_input_text.bxml"
test case.