Added a new "expand(Insets)" method to Dimensions, as this is a fairly common thing to do.  Added tests to DimensionsTest also.
Some cleaning up in TerraTooltipSkin both to use this new method
but also to simplify the fact that there were a bunch of +2 and +1
to the padding values, which seems silly.  So just increased the
original (only) padding values, and make more room on left/right
so the text is more uniformly padded all around.