PIVOT-1031:  Add Gauge to the ComponentExplorer under "Progress Indicators" (just like Meter and ActivityIndicator, as in the Kitchen Sink tutorial).
But, because Gauge is a parameterized control, some changes had to be made
in the ComponentExplorer and elsewhere to support it:
* Update BeanMonitor to deal with the parameterized listener interface.
* Also same changes to EventLogger for the parameterized listener.
* Made a change to ComponentInspectorSkin to add a label for Font and Number
  types to at least display their values, even if they can't really be changed.

But because of the way "notifying properties" are determined, had to rework the
GaugeListener to have individual methods for min, max, warning, and critical
values so they would show up in the property inspector.  Change all the users.

Add the "gauge.bxml" and the tree menu item for it in the ComponentExplorer
itself.