The "Style" enum was introduced to make setting component styles more foolproof by defining the style names in terms of an enum that can be checked at compile-time.  So, go through all the uses of "getStyles" and correct anything that is still being set via a string constant (if possible, some things are just custom JSON, for instance).
Add a few Style values that were being set via strings.

Make some "check-style" changes in the files that were changed,
as well as convert some callbacks to lambdas as appropriate.

A bunch of changes to TextPaneDemo suggested some improvements
to Span, which then required new tests in SpanTest as well. And
a new method added to Node: "getParagraph".